server/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
diskcache==5.6.3
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
llama_cpp_python==0.2.89
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.1
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
regex==2024.7.24
requests==2.32.3
setuptools==73.0.1
sympy==1.13.2
tiktoken==0.7.0
torch==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
#ForEmail
smtplib
ssl
email
rapidfuzz==3.1.1